requirements.txt
paho-mqtt==1.5.0
jsonschema==3.2.0
